LayoutTests/platform/blackberry/fast/filesystem/file-entry-to-uri-expected.txt

CONSOLE MESSAGE: line 33: ReferenceError: Can't find variable: webkitRequestFileSystem
FAIL: Timed out waiting for notifyDone to be called
Obtaining URL from FileEntry.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
